document pr_json model, small fixes
